refactoring to add a static filtering and ordering method in PluginLocatorUtils
